requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
motor~=3.3.2
pydantic~=2.5.3
pymongo~=4.6.1
starlette~=0.27.0
jose~=1.0.0